Tried to configure project for windows and vs2012

+++ CellarWorkbench/Graphics/GL/GlShaderBank.cpp
@@ -94,7 +94,7 @@ namespace cellar
             const std::string& shaderName,
             GLenum)
     {
-        int nbRemoved = _shaders.erase(shaderName);
+		size_t nbRemoved = _shaders.erase(shaderName);
 
         if(nbRemoved)
         {
